third_party/WebKit/LayoutTests/editing/caret/caret-direction-auto-expected.txt

Tests various scenarios for caret position when direction is auto

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textarea.value is "א!"
1. Caret is at right if text is in LTR direction
PASS textarea.value is "א!\nhello"
Compares lines if they are different
PASS beforeCaretPos[1] is not afterCaretPos[1]
New text is in LTR direction, start of caret will differ
PASS beforeCaretPos[0] is > afterCaretPos[0]
2. Addition of RTL text in middle of LTR text, caret moves to the left of the line.
PASS textarea.value is "א!\nhelloאא"
RTL text added will be move caret to start of position where hello begins
PASS beforeCaretPos[0] is afterCaretPos[0]+afterCaretPos[2]
3. Adding LTR text after RTL should move caret towards right.
PASS textarea.value is "א!\nhelloאאbye"
Start position of caret is same
PASS beforeCaretPos[0]+beforeCaretPos[2] is afterCaretPos[0]
But end position differs
PASS beforeCaretPos[0]+beforeCaretPos[2] is not afterCaretPos[0]+afterCaretPos[2]
4. Adding space, caret should move caret
PASS textarea.value is "א!\nhelloאאbye "
PASS afterCaretPos[0]+afterCaretPos[2] is > beforeCaretPos[0]+beforeCaretPos[2]
PASS successfullyParsed is true

TEST COMPLETE
